Create notebook to plot data from new RawDataset. Also use float64 for datetimes. Closes #84

+++ power_perceiver/utils.py
@@ -7,7 +7,7 @@ import pandas as pd
 from pathy import Pathy
 
 
-def datetime64_to_float(datetimes: np.ndarray, dtype=np.float32) -> np.ndarray:
+def datetime64_to_float(datetimes: np.ndarray, dtype=np.float64) -> np.ndarray:
     nums = datetimes.astype("datetime64[s]").astype(dtype)
     mask = np.isfinite(datetimes)
     return np.where(mask, nums, np.NaN)
